refactor: TracingConfig, SentryConfig and MetricsConfig
