+++ src/infra/mod.rs
@@ -8,6 +8,3 @@ pub mod sentry;
 pub mod tracing;
 
 pub use blockchain_client::BlockchainClient;
-pub use metrics::init_metrics;
-pub use sentry::init_sentry;
-pub use tracing::init_tracing;
