+++ src/infra/sentry/mod.rs
@@ -1,0 +1,3 @@
+mod sentry_config;
+
+pub use sentry_config::SentryConfig;
