+++ src/infra/tracing/mod.rs
@@ -1,15 +1,16 @@
-#[allow(clippy::module_inception)]
-mod tracing;
 mod tracing_config;
+mod tracing_services;
 
-pub use tracing::info_task_spawn;
-pub use tracing::init_tracing;
-pub use tracing::new_cid;
-pub use tracing::warn_task_cancellation;
-pub use tracing::warn_task_rx_closed;
-pub use tracing::warn_task_tx_closed;
-pub use tracing::SpanExt;
-pub use tracing::TracingExt;
 pub use tracing_config::TracingConfig;
 pub use tracing_config::TracingLogFormat;
 pub use tracing_config::TracingProtocol;
+pub use tracing_services::info_task_spawn;
+pub use tracing_services::new_cid;
+pub use tracing_services::warn_task_cancellation;
+pub use tracing_services::warn_task_rx_closed;
+pub use tracing_services::warn_task_tx_closed;
+pub use tracing_services::SpanExt;
+pub use tracing_services::TracingContextLayer;
+pub use tracing_services::TracingExt;
+pub use tracing_services::TracingJsonFormatter;
+pub use tracing_services::TracingMinimalTimer;
